fix(settings): restore missing ui/AccordionSection module
SettingsOverlay.tsx imported Disclosure/DisclosureChevron from
./ui/AccordionSection, but that file was never committed — breaking
the dev build (TS2307). IntelligenceSettings.tsx had its own local
copies of the same two components (also importing from the missing
file, which would collide as a duplicate declaration once the file
existed).

Extract Disclosure/DisclosureChevron into src/components/ui/AccordionSection.tsx
and have both files import the shared component instead of duplicating it.
